# convert playcount.Watched to map