fixes https://github.com/elgatito/plugin.video.elementum/issues/796

but in reality speed gain is minimal. even with 50x speed-up it is still µs
old:
```
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Futurama took 7.575µs
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for South Park took 463.038µs
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Firefly took 28.041µs
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Family Guy took 511.418µs
```

new:
```
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Futurama took 375ns
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for South Park took 10.463µs
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Firefly took 379ns
[plugin.video.elementum] DEBU  tmdb         ▶ ToListItem       watchedEpisodesNumber for Family Guy took 10.416µs
```